refactor: notificationlog -> meetingRoom 패키지명 변경
