+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/BindingAdapter.kt
@@ -1,10 +1,10 @@
-package com.woowacourse.ody.presentation.notificationlog
+package com.woowacourse.ody.presentation.meetingroom
 
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import com.woowacourse.ody.R
 import com.woowacourse.ody.domain.model.NotificationType
-import com.woowacourse.ody.presentation.notificationlog.uimodel.NotificationLogUiModel
+import com.woowacourse.ody.presentation.meetingroom.uimodel.NotificationLogUiModel
 
 @BindingAdapter("setNotificationLogText")
 fun TextView.setNotificationLogText(log: NotificationLogUiModel) {
