+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/CopyInviteCodeListener.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.notificationlog
+package com.woowacourse.ody.presentation.meetingroom
 
 interface CopyInviteCodeListener {
     fun copyInviteCode()
