+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/MeetingRoomViewModelFactory.kt
@@ -1,17 +1,17 @@
-package com.woowacourse.ody.presentation.notificationlog
+package com.woowacourse.ody.presentation.meetingroom
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.woowacourse.ody.domain.repository.ody.MeetingRepository
 import com.woowacourse.ody.domain.repository.ody.NotificationLogRepository
 
-class NotificationLogViewModelFactory(
+class MeetingRoomViewModelFactory(
     private val notificationLogRepository: NotificationLogRepository,
     private val meetingRepository: MeetingRepository,
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return if (modelClass.isAssignableFrom(NotificationLogViewModel::class.java)) {
-            NotificationLogViewModel(
+        return if (modelClass.isAssignableFrom(MeetingRoomViewModel::class.java)) {
+            MeetingRoomViewModel(
                 notificationLogRepository,
                 meetingRepository,
             ) as T
