+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/ShareListener.kt
@@ -1,0 +1,5 @@
+package com.woowacourse.ody.presentation.meetingroom
+
+interface ShareListener {
+    fun share()
+}
