+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/adapter/InviteCodeViewHolder.kt
@@ -1,8 +1,8 @@
-package com.woowacourse.ody.presentation.notificationlog.adapter
+package com.woowacourse.ody.presentation.meetingroom.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.woowacourse.ody.databinding.ItemInviteCodeLogBinding
-import com.woowacourse.ody.presentation.notificationlog.CopyInviteCodeListener
+import com.woowacourse.ody.presentation.meetingroom.CopyInviteCodeListener
 
 class InviteCodeViewHolder(private val binding: ItemInviteCodeLogBinding) :
     RecyclerView.ViewHolder(binding.root) {
