+++ android/app/src/main/res/layout/bottom_sheet_code_copy.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="copyInviteCodeListener"
-            type="com.woowacourse.ody.presentation.notificationlog.CopyInviteCodeListener" />
+            type="com.woowacourse.ody.presentation.meetingroom.CopyInviteCodeListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
