+++ android/app/src/main/res/layout/item_invite_code_log.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="listener"
-            type="com.woowacourse.ody.presentation.notificationlog.CopyInviteCodeListener" />
+            type="com.woowacourse.ody.presentation.meetingroom.CopyInviteCodeListener" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
